pattern.py => aml/: Split pattern.py into smaller files.
