+++ generation/src/proof_generation/aml/__init__.py
@@ -1,0 +1,4 @@
+from .matching import *
+from .notation import *
+from .notation import _and, _or
+from .syntax import *
