+++ generation/src/proof_generation/aml/syntax.py
@@ -9,74 +9,6 @@ if TYPE_CHECKING:
     from collections.abc import Mapping
 
 
-def match_single(
-    pattern: Pattern, instance: Pattern, extend: dict[int, Pattern] | None = None
-) -> dict[int, Pattern] | None:
-    ret: dict[int, Pattern] | None
-    ret = extend if extend else {}
-
-    if isinstance(pattern, MetaVar):
-        id = pattern.name
-        if id in ret:
-            if ret[id] != instance:
-                return None
-        else:
-            if not pattern.can_be_replaced_by(instance):
-                return None
-            ret[id] = instance
-        return ret
-    if (pat_imp := Implies.unwrap(pattern)) and (inst_imp := Implies.unwrap(instance)):
-        ret = match_single(pat_imp[0], inst_imp[0], ret)
-        if ret is None:
-            return None
-        return match_single(pat_imp[1], inst_imp[1], ret)
-    # The following three cases are more verbose because a 0 returned by the
-    # deconstruct is still interpreted as False even though it is not None
-    pat_evar = EVar.deconstruct(pattern)
-    inst_evar = EVar.deconstruct(instance)
-    if (pat_evar is not None) and (inst_evar is not None):
-        if pat_evar != inst_evar:
-            return None
-        return ret
-    pat_svar = SVar.deconstruct(pattern)
-    inst_svar = SVar.deconstruct(instance)
-    if (pat_svar is not None) and (inst_svar is not None):
-        if pat_svar != inst_svar:
-            return None
-        return ret
-    pat_sym = Symbol.deconstruct(pattern)
-    inst_sym = Symbol.deconstruct(instance)
-    if (pat_sym is not None) and (inst_sym is not None):
-        if pat_sym != inst_sym:
-            return None
-        return ret
-    if (pat_app := App.unwrap(pattern)) and (inst_app := App.unwrap(instance)):
-        ret = match_single(pat_app[0], inst_app[0], ret)
-        if ret is None:
-            return None
-        return match_single(pat_app[1], inst_app[1], ret)
-    if (pat_ex := Exists.deconstruct(pattern)) and (inst_ex := Exists.deconstruct(instance)):
-        if pat_ex[0] != inst_ex[0]:
-            return None
-        return match_single(pat_ex[1], inst_ex[1], ret)
-    if (pat_mu := Mu.deconstruct(pattern)) and (inst_mu := Mu.deconstruct(instance)):
-        if pat_mu[0] != inst_mu[0]:
-            return None
-        return match_single(pat_mu[1], inst_mu[1], ret)
-    # TODO Consider adding cases for ESubst/SSubst
-    return None
-
-
-def match(equations: list[tuple[Pattern, Pattern]]) -> dict[int, Pattern] | None:
-    ret: dict[int, Pattern] = {}
-    for pattern, instance in equations:
-        submatch = match_single(pattern, instance, ret)
-        if not submatch:
-            return None
-        ret = submatch
-    return ret
-
-
 class Pattern:
     def evar_is_fresh(self, name: int) -> bool:
         raise NotImplementedError
@@ -618,53 +550,8 @@ class Instantiate(Pattern):
 
 
 @dataclass(frozen=True)
-class Notation:
-    label: str
-    arity: int
-    definition: Pattern
-    format_str: str
-
-    def __post_init__(self) -> None:
-        if self.definition.metavars():
-            assert (
-                max(self.definition.metavars()) < self.arity
-            ), f'Notation {self.label}: Number of variables used is greater than Arity.'
-
-        assert self.definition.occurring_vars() == set()
-
-    def __call__(self, *args: Pattern) -> Pattern:
-        assert len(args) == self.arity, f'Notation {self.label}: expected {self.arity} arguements, got {len(args)}.'
-        return Instantiate(self.definition, frozendict(enumerate(args)))
-
-    def matches(self, pattern: Pattern) -> None | tuple[Pattern, ...]:
-        match = match_single(self.definition, pattern)
-        if match is None:
-            return None
-        return tuple((match[i] if i in match else MetaVar(i)) for i in range(self.arity))
-
-    def assert_matches(self, pattern: Pattern) -> tuple[Pattern, ...]:
-        if match := self.matches(pattern):
-            return match
-        raise AssertionError(f'Does not match notation {self.label}: {str(pattern)}')
-
-    def print_instantiation(self, applied: Instantiate, opts: PrettyOptions) -> str:
-        assert applied.pattern == self.definition
-        pretty_opts = [p.pretty(opts) for _, p in sorted(applied.inst.items())]
-        try:
-            return self.format_str.format(*pretty_opts)
-        except Exception as e:
-            raise ValueError(f'Cannot format malformed notation {self.label}: {self.format_str}') from e
-
-
-@dataclass(frozen=True)
 class PrettyOptions:
     simplify_instantiations: bool = False
     notations: Mapping[Pattern, Notation] = frozendict({})
 
 
-bot = Notation('bot', 0, Mu(0, SVar(0)), '⊥')
-neg = Notation('not', 1, Implies(MetaVar(0), bot()), '¬{0}')
-top = Notation('top', 0, neg(bot()), '⊤')
-_and = Notation('and', 2, neg(Implies(phi0, neg(phi1))), '({0} ⋀ {1})')
-_or = Notation('or', 2, Implies(neg(phi0), phi1), '({0} ⋁ {1})')
-equiv = Notation('equiv', 2, _and(Implies(phi0, phi1), Implies(phi1, phi0)), f'({0} <-> {1})')
