+++ generation/src/proof_generation/interpreter.py
@@ -8,9 +8,9 @@ if TYPE_CHECKING:
     from collections.abc import Mapping
 
     from .proved import Proved
-    from proof_generation.pattern import Pattern
+    from .aml import Pattern
 
-from proof_generation.pattern import App, ESubst, EVar, Exists, Implies, Instantiate, MetaVar, Mu, SSubst, SVar, Symbol
+from proof_generation.aml import App, ESubst, EVar, Exists, Implies, Instantiate, MetaVar, Mu, SSubst, SVar, Symbol
 
 
 class ExecutionPhase(Enum):
