+++ generation/src/proof_generation/stateful_interpreter.py
@@ -8,9 +8,9 @@ from proof_generation.proved import Proved
 if TYPE_CHECKING:
     from collections.abc import Mapping
 
+    from proof_generation.aml import ESubst, EVar, MetaVar, Pattern, SSubst, SVar
     from proof_generation.claim import Claim
     from proof_generation.interpreter import ExecutionPhase
-    from proof_generation.pattern import ESubst, EVar, MetaVar, Pattern, SSubst, SVar
 
 
 class StatefulInterpreter(BasicInterpreter):
